refractor Apollo Server code to
